fix warnings in script.h

+++ arm9/src/script.c
@@ -1,4 +1,3 @@
-#include "global.h"
 #include "script.h"
 
 u16 ScriptReadHalfword(struct ScriptContext *ctx);
@@ -42,9 +41,9 @@ THUMB_FUNC void StopScript(struct ScriptContext *ctx)
     ctx->scriptPtr = 0;
 }
 
-THUMB_FUNC void FUN_02038B6C(struct ScriptContext *ctx, int r1)
+THUMB_FUNC void FUN_02038B6C(struct ScriptContext *ctx, s32 r1)
 {
-    ctx->unk74 = r1;
+    ctx->unk74 = (u32)r1;
 }
 
 THUMB_FUNC u8 RunScriptCommand(struct ScriptContext *ctx)
